Fix warning when enabling the updater
